feat: Lots of changes to enable better testing
Initial work on integration testing to confirm ADF compatibility
