--- esbuild-cli.config.mjs
@@ -1,31 +1,0 @@
-import esbuild from "esbuild";
-import process from "process";
-import builtins from 'builtin-modules'
-
-const banner =
-`/*
-THIS IS A GENERATED/BUNDLED FILE BY ESBUILD
-if you want to view the source, please visit the github repository of this plugin
-*/
-`;
-
-const prod = (process.argv[2] === 'production');
-
-esbuild.build({
-	banner: {
-		js: banner,
-	},
-	entryPoints: ['src/cli.ts'],
-	bundle: true,
-	format: 'esm',
-	watch: false,
-	target: 'node19',
-	platform: 'node',
-	logLevel: "info",
-	sourcemap: prod ? false : 'inline',
-	treeShaking: false,
-	outfile: 'dist-cli/cli.js',
-	mainFields: ['module', 'main'],
-	external: [
-		...builtins],
-}).catch(() => process.exit(1));
