+++ esbuild.config.mjs
@@ -2,6 +2,7 @@ import esbuild from "esbuild";
 import process from "process";
 import builtins from 'builtin-modules'
 import { writeFileSync } from 'fs';
+import { mermaidRendererPlugin } from './.build/mermaidRendererPlugin.js'
 
 
 const banner =
@@ -10,55 +11,6 @@ THIS IS A GENERATED/BUNDLED FILE BY ESBUILD
 if you want to view the source, please visit the github repository of this plugin
 */
 `;
-
-
-const mermaid_renderer_plugin = {
-	name: 'mermaid_renderer_plugin',
-	setup(build) {
-		build.onResolve({ filter: /mermaid_renderer\.esbuild$/ }, args => {
-			return {
-				watchFiles: ['src/mermaid_renderer.js'],
-				path: args.path,
-				namespace: 'mermaid-binary',
-			  }
-		});
-		build.onLoad({ filter: /mermaid_renderer\.esbuild$/, namespace: 'mermaid-binary' }, async (args) => {
-			const result = await esbuild.build({
-				entryPoints: ['src/mermaid_renderer.js'],
-				bundle: true,
-				format: 'cjs',
-				target: 'chrome106',
-				logLevel: 'info',
-				sourcemap: false,
-				treeShaking: true,
-				write: false,
-				mainFields: ['module', 'main'],
-				minify: true,
-			}).catch(() => process.exit(1));
-
-			const fileContents = `
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8" />
-    <title>Mermaid Chart</title>
-  </head>
-  <body>
-  	<div id="graphDiv"></div>
-    <script type="text/javascript">
-	${result.outputFiles[0].text}
-	</script>
-  </body>
-</html>
-			`;
-
-			return {
-				contents: fileContents,
-				loader: 'binary',
-			}
-		})
-	}
-};
 
 const prod = (process.argv[2] === 'production');
 
@@ -91,7 +43,7 @@ const buildConfig = {
 	treeShaking: true,
 	outdir: prod ? 'dist' : 'dev-vault/.obsidian/plugins/obsidian-confluence',
 	mainFields: ['module', 'main'],
-	plugins: [mermaid_renderer_plugin],
+	plugins: [mermaidRendererPlugin],
 	minify: true,
 	metafile: true,
 };
