+++ src/Publisher.ts
@@ -2,29 +2,34 @@ import { ConfluenceSettings } from "./Settings";
 import FolderFile from "./FolderFile.json";
 
 import { traverse, filter } from "@atlaskit/adf-utils/traverse";
-import SparkMD5 from "spark-md5";
-import { doc, p } from "@atlaskit/adf-utils/builders";
 import { ADFEntity } from "@atlaskit/adf-utils/types";
-import MdToADF from "./mdToADF";
-import {
-	ConfluenceClient,
-	LoaderAdaptor,
-	MarkdownFile,
-} from "./adaptors/types";
-import { MermaidRenderer, ChartData } from "./mermaid_renderers/types";
-import * as path from "path";
+import { CustomConfluenceClient, LoaderAdaptor } from "./adaptors/types";
 import { JSONDocNode } from "@atlaskit/editor-json-transformer";
 import { UploadResults } from "./CompletedView";
+import {
+	getMermaidFileName,
+	MermaidRenderer,
+	ChartData,
+} from "./mermaid_renderers";
+import { createFolderStructure as createLocalAdfTree } from "./TreeLocal";
+import { ensureAllFilesExistInConfluence } from "./TreeConfluence";
+import { uploadBuffer, UploadedImageData, uploadFile } from "./Attachments";
+import { prepareAdf } from "./AdfPostProcess";
+import deepEqual from "deep-equal";
 
-const mdToADFConverter = new MdToADF();
+export interface LocalAdfFileTreeNode {
+	name: string;
+	children: LocalAdfFileTreeNode[];
+	file?: LocalAdfFile;
+}
 
 interface FilePublishResult {
 	successfulUpload: boolean;
-	absoluteFilePath: string;
+	node: ConfluenceNode;
 	reason?: string;
 }
 
-export interface AdfFile {
+export interface LocalAdfFile {
 	folderName: string;
 	absoluteFilePath: string;
 	fileName: string;
@@ -36,45 +41,35 @@ export interface AdfFile {
 	tags: string[];
 	pageId: string | undefined;
 	dontChangeParentPageId: boolean;
-	spaceKey?: string;
 }
 
-interface ConfluenceNode {
-	file: AdfFile;
+export interface ConfluenceAdfFile {
+	folderName: string;
+	absoluteFilePath: string;
+	fileName: string;
+	contents: JSONDocNode;
+	pageTitle: string;
+	frontmatter: {
+		[key: string]: unknown;
+	};
+	tags: string[];
+	dontChangeParentPageId: boolean;
+
+	pageId: string;
+	spaceKey: string;
+}
+
+export interface ConfluenceNode {
+	file: ConfluenceAdfFile;
 	version: number;
 	existingAdf: string;
 	parentPageId: string;
 }
-interface ConfluenceTreeNode {
-	file: AdfFile;
+export interface ConfluenceTreeNode {
+	file: ConfluenceAdfFile;
 	version: number;
 	existingAdf: string;
 	children: ConfluenceTreeNode[];
-}
-
-function flattenTree(
-	node: ConfluenceTreeNode,
-	parentPageId?: string
-): ConfluenceNode[] {
-	const nodes: ConfluenceNode[] = [];
-	const { file, version, existingAdf, children } = node;
-
-	if (parentPageId) {
-		nodes.push({
-			file,
-			version,
-			existingAdf,
-			parentPageId: parentPageId,
-		});
-	}
-
-	if (children) {
-		children.forEach((child) => {
-			nodes.push(...flattenTree(child, file.pageId));
-		});
-	}
-
-	return nodes;
 }
 
 export class Publisher {
@@ -97,7 +92,7 @@ export class Publisher {
 		this.confluenceClient = confluenceClient;
 	}
 
-	async doPublish(publishFilter?: string): Promise<UploadResults> {
+	async publish(publishFilter?: string) {
 		const parentPage = await this.confluenceClient.content.getContentById({
 			id: this.settings.confluenceParentId,
 			expand: ["body.atlas_doc_format", "space"],
@@ -109,65 +104,17 @@ export class Publisher {
 		const spaceToPublishTo = parentPage.space;
 
 		const files = await this.adaptor.getMarkdownFilesToUpload();
-		const folderTree = createFolderStructure(files);
-		const confluencePageTree = await this.createFileStructureInConfluence(
+		const folderTree = createLocalAdfTree(files);
+		let confluencePagesToPublish = await ensureAllFilesExistInConfluence(
+			this.confluenceClient,
+			this.adaptor,
 			folderTree,
 			spaceToPublishTo.key,
 			parentPage.id,
-			parentPage.id,
-			false
+			parentPage.id
 		);
-		let confluencePagesToPublish = flattenTree(confluencePageTree);
 
-		const fileToPageIdMap: Record<string, AdfFile> = {};
-
-		confluencePagesToPublish.forEach((node) => {
-			if (!node.file.pageId) {
-				throw new Error("Missing Page ID");
-			}
-
-			fileToPageIdMap[node.file.fileName] = node.file;
-		});
-
-		confluencePagesToPublish.forEach((node) => {
-			node.file.contents = traverse(node.file.contents, {
-				text: (node, _parent) => {
-					if (
-						node.marks &&
-						node.marks[0].type === "link" &&
-						node.marks[0].attrs
-					) {
-						if (
-							typeof node.marks[0].attrs.href === "string" &&
-							node.marks[0].attrs.href.startsWith("wikilink")
-						) {
-							const wikilinkUrl = new URL(
-								node.marks[0].attrs.href
-							);
-							const pagename = `${wikilinkUrl.pathname}.md`;
-
-							const linkPage = fileToPageIdMap[pagename];
-							if (linkPage) {
-								const confluenceUrl = `${this.settings.confluenceBaseUrl}/wiki/spaces/${linkPage.spaceKey}/pages/${linkPage.pageId}${wikilinkUrl.hash}`;
-								node.marks[0].attrs.href = confluenceUrl;
-								if (node.text === wikilinkUrl.pathname) {
-									node.type = "inlineCard";
-									node.attrs = {
-										url: node.marks[0].attrs.href,
-									};
-									delete node.marks;
-									delete node.text;
-									return node;
-								}
-							} else {
-								node.marks.remove(node.marks[0]);
-							}
-							return node;
-						}
-					}
-				},
-			}) as JSONDocNode;
-		});
+		prepareAdf(confluencePagesToPublish, this.settings);
 
 		if (publishFilter) {
 			confluencePagesToPublish = confluencePagesToPublish.filter(
@@ -180,6 +127,11 @@ export class Publisher {
 		});
 
 		const adrFiles = await Promise.all(adrFileTasks);
+		return adrFiles;
+	}
+
+	async doPublish(publishFilter?: string): Promise<UploadResults> {
+		const adrFiles = await this.publish(publishFilter);
 
 		const returnVal: UploadResults = {
 			successfulUploads: 0,
@@ -194,7 +146,7 @@ export class Publisher {
 			}
 
 			returnVal.failedFiles.push({
-				fileName: element.absoluteFilePath,
+				fileName: element.node.file.absoluteFilePath,
 				reason: element.reason ?? "No Reason Provided",
 			});
 		});
@@ -202,211 +154,40 @@ export class Publisher {
 		return returnVal;
 	}
 
-	async createFileStructureInConfluence(
-		node: TreeNode,
-		spaceKey: string,
-		parentPageId: string,
-		topPageId: string,
-		createPage: boolean
-	): Promise<ConfluenceTreeNode> {
-		if (!node.file) {
-			throw new Error("Missing file on node");
-		}
-
-		let version: number;
-		let existingAdf: string | undefined;
-
-		if (createPage) {
-			const pageDetails = await this.ensurePageExists(
-				node.file,
-				spaceKey,
-				parentPageId,
-				topPageId
-			);
-			node.file.pageId = pageDetails.id;
-			node.file.spaceKey = pageDetails.spaceKey;
-			version = pageDetails.version;
-			existingAdf = pageDetails.existingAdf;
-		} else {
-			node.file.pageId = parentPageId;
-			node.file.spaceKey = spaceKey;
-			version = 0;
-			existingAdf = "";
-		}
-
-		if (node.file.pageId === undefined) {
-			throw new Error("Missing myPageId");
-		}
-
-		const nonNullMyPageId = node.file.pageId;
-
-		const childDetailsTasks = node.children.map((childNode) => {
-			return this.createFileStructureInConfluence(
-				childNode,
-				spaceKey,
-				nonNullMyPageId,
-				topPageId,
-				true
-			);
-		});
-
-		const childDetails = await Promise.all(childDetailsTasks);
-
-		return {
-			file: node.file,
-			version,
-			existingAdf: existingAdf ?? "",
-			children: childDetails,
-		};
-	}
-
-	async ensurePageExists(
-		file: AdfFile,
-		spaceKey: string,
-		parentPageId: string,
-		topPageId: string
-	) {
-		if (file.pageId) {
-			const contentById =
-				await this.confluenceClient.content.getContentById({
-					id: file.pageId,
-					expand: [
-						"version",
-						"body.atlas_doc_format",
-						"ancestors",
-						"space",
-					],
-				});
-
-			if (!contentById.space?.key) {
-				throw new Error("Missing Space Key");
-			}
-
-			return {
-				id: contentById.id,
-				title: file.pageTitle,
-				version: contentById?.version?.number ?? 1,
-				existingAdf: contentById?.body?.atlas_doc_format?.value,
-				spaceKey: contentById.space.key,
-			};
-		}
-
-		const searchParams = {
-			type: "page",
-			spaceKey,
-			title: file.pageTitle,
-			expand: ["version", "body.atlas_doc_format", "ancestors"],
-		};
-		const contentByTitle = await this.confluenceClient.content.getContent(
-			searchParams
-		);
-
-		if (contentByTitle.size > 0) {
-			const currentPage = contentByTitle.results[0];
-
-			// TODO: Ensure is part of topPageId page tree
-			console.log({ ancestors: currentPage.ancestors });
-			if (
-				!currentPage.ancestors?.some(
-					(ancestor) => ancestor.id == topPageId
-				)
-			) {
-				throw new Error(
-					`${file.pageTitle} is trying to overwrite a page outside the page tree from the selected top page`
-				);
-			}
-
-			await this.adaptor.updateMarkdownPageId(
-				file.absoluteFilePath,
-				currentPage.id
-			);
-			return {
-				id: currentPage.id,
-				title: file.pageTitle,
-				version: currentPage?.version?.number ?? 1,
-				existingAdf: currentPage?.body?.atlas_doc_format?.value,
-				spaceKey,
-			};
-		} else {
-			console.log("Creating page");
-			const creatingBlankPageRequest = {
-				space: { key: spaceKey },
-				ancestors: [{ id: parentPageId }],
-				title: file.pageTitle,
-				type: "page",
-				body: {
-					// eslint-disable-next-line @typescript-eslint/naming-convention
-					atlas_doc_format: {
-						value: this.blankPageAdf,
-						representation: "atlas_doc_format",
-					},
-				},
-			};
-			const pageDetails =
-				await this.confluenceClient.content.createContent(
-					creatingBlankPageRequest
-				);
-
-			await this.adaptor.updateMarkdownPageId(
-				file.absoluteFilePath,
-				pageDetails.id
-			);
-			return {
-				id: pageDetails.id,
-				title: file.pageTitle,
-				version: pageDetails?.version?.number ?? 1,
-				existingAdf: pageDetails?.body?.atlas_doc_format?.value,
-				spaceKey,
-			};
-		}
-	}
-
-	async publishFile(node: ConfluenceNode): Promise<FilePublishResult> {
+	private async publishFile(
+		node: ConfluenceNode
+	): Promise<FilePublishResult> {
 		try {
-			if (!node.file.pageId) {
-				throw new Error("Missing Page ID");
-			}
-
 			await this.updatePageContent(
-				node.file.pageId,
-				node.file.absoluteFilePath,
-				node.file.contents,
 				node.parentPageId,
 				node.version,
-				node.file.pageTitle,
 				node.existingAdf,
-				node.file.tags,
 				node.file
 			);
 
 			return {
 				successfulUpload: true,
-				absoluteFilePath: node.file.absoluteFilePath,
+				node,
 			};
 		} catch (e: unknown) {
 			return {
 				successfulUpload: false,
-				absoluteFilePath: node.file.absoluteFilePath,
-				reason: JSON.stringify(e),
+				node,
+				reason: JSON.stringify(e), // TODO: Understand why this doesn't show error message properly
 			};
 		}
 	}
 
-	async updatePageContent(
-		pageId: string,
-		originFileAbsoluteFilePath: string,
-		adf: ADFEntity,
+	private async updatePageContent(
 		parentPageId: string,
 		pageVersionNumber: number,
-		pageTitle: string,
 		currentContents: string,
-		labels: string[],
-		adfFile: AdfFile
+		adfFile: ConfluenceAdfFile
 	) {
 		const updatedAdf = await this.uploadFiles(
-			pageId,
-			originFileAbsoluteFilePath,
-			adf
+			adfFile.pageId,
+			adfFile.absoluteFilePath,
+			adfFile.contents
 		);
 
 		const adr = JSON.stringify(updatedAdf);
@@ -415,18 +196,18 @@ export class Publisher {
 		console.log(currentContents);
 		console.log(adr);
 
-		if (currentContents === adr) {
+		if (deepEqual(JSON.parse(currentContents), updatedAdf)) {
 			console.log("Page is the same not updating");
 			return true;
 		}
 
 		const updateContentDetails = {
-			id: pageId,
+			id: adfFile.pageId,
 			...(adfFile.dontChangeParentPageId
 				? {}
 				: { ancestors: [{ id: parentPageId }] }),
 			version: { number: pageVersionNumber + 1 },
-			title: pageTitle,
+			title: adfFile.pageTitle,
 			type: "page",
 			body: {
 				// eslint-disable-next-line @typescript-eslint/naming-convention
@@ -436,24 +217,20 @@ export class Publisher {
 				},
 			},
 		};
-		console.log({ updateContentDetails });
 		await this.confluenceClient.content.updateContent(updateContentDetails);
 		const getLabelsForContent = {
-			id: pageId,
+			id: adfFile.pageId,
 		};
 		const currentLabels =
 			await this.confluenceClient.contentLabels.getLabelsForContent(
 				getLabelsForContent
 			);
-		if (currentLabels.size > 0) {
-			console.log({ currentLabels });
-		}
 
 		for (const existingLabel of currentLabels.results) {
-			if (!labels.includes(existingLabel.label)) {
+			if (!adfFile.tags.includes(existingLabel.label)) {
 				await this.confluenceClient.contentLabels.removeLabelFromContentUsingQueryParameter(
 					{
-						id: pageId,
+						id: adfFile.pageId,
 						name: existingLabel.name,
 					}
 				);
@@ -461,7 +238,7 @@ export class Publisher {
 		}
 
 		const labelsToAdd = [];
-		for (const newLabel of labels) {
+		for (const newLabel of adfFile.tags) {
 			if (
 				currentLabels.results.findIndex(
 					(item) => item.label === newLabel
@@ -476,13 +253,13 @@ export class Publisher {
 
 		if (labelsToAdd.length > 0) {
 			await this.confluenceClient.contentLabels.addLabelsToContent({
-				id: pageId,
+				id: adfFile.pageId,
 				body: labelsToAdd,
 			});
 		}
 	}
 
-	async uploadFiles(
+	private async uploadFiles(
 		pageId: string,
 		pageFilePath: string,
 		adr: ADFEntity
@@ -517,6 +294,14 @@ export class Publisher {
 				...mermaidNodesToUpload,
 			]);
 
+		const imagesToUpload = new Set(
+			mediaNodes.map((node) => node?.attrs?.url)
+		);
+
+		if (imagesToUpload.size === 0 && mermaidChartsAsImages.size === 0) {
+			return adr;
+		}
+
 		const currentUploadedAttachments =
 			await this.confluenceClient.contentAttachments.getAttachments({
 				id: pageId,
@@ -536,16 +321,13 @@ export class Publisher {
 			{}
 		);
 
-		const imagesToUpload = new Set(
-			mediaNodes.map((node) => node?.attrs?.url)
-		).values();
-
 		let imageMap: Record<string, UploadedImageData | null> = {};
 
-		for (const imageUrl of imagesToUpload) {
-			console.log({ testing: imageUrl });
+		for (const imageUrl of imagesToUpload.values()) {
 			const filename = imageUrl.split("://")[1];
-			const uploadedContent = await this.uploadFile(
+			const uploadedContent = await uploadFile(
+				this.confluenceClient,
+				this.adaptor,
 				pageId,
 				pageFilePath,
 				filename,
@@ -559,7 +341,8 @@ export class Publisher {
 		}
 
 		for (const mermaidImage of mermaidChartsAsImages) {
-			const uploadedContent = await this.uploadBuffer(
+			const uploadedContent = await uploadBuffer(
+				this.confluenceClient,
 				pageId,
 				mermaidImage[0],
 				mermaidImage[1],
@@ -575,7 +358,6 @@ export class Publisher {
 		const afterAdf = traverse(adr, {
 			media: (node, _parent) => {
 				if (node?.attrs?.type === "file") {
-					console.log({ node });
 					if (!imageMap[node?.attrs?.url]) {
 						return;
 					}
@@ -584,7 +366,6 @@ export class Publisher {
 						node.attrs.collection = mappedImage.collection;
 						node.attrs.id = mappedImage.id;
 						delete node.attrs.url;
-						console.log({ node });
 						return node;
 					}
 				}
@@ -626,230 +407,4 @@ export class Publisher {
 
 		return afterAdf;
 	}
-
-	async uploadBuffer(
-		pageId: string,
-		uploadFilename: string,
-		fileBuffer: Buffer,
-		currentAttachments: Record<
-			string,
-			{ filehash: string; attachmentId: string; collectionName: string }
-		>
-	): Promise<UploadedImageData | null> {
-		const spark = new SparkMD5.ArrayBuffer();
-		const currentFileMd5 = spark.append(fileBuffer).end();
-
-		if (
-			!!currentAttachments[uploadFilename] &&
-			currentAttachments[uploadFilename].filehash === currentFileMd5
-		) {
-			return {
-				filename: uploadFilename,
-				id: currentAttachments[uploadFilename].attachmentId,
-				collection: currentAttachments[uploadFilename].collectionName,
-			};
-		}
-
-		const attachmentDetails = {
-			id: pageId,
-			attachments: [
-				{
-					file: fileBuffer,
-					filename: uploadFilename,
-					minorEdit: false,
-					comment: currentFileMd5,
-					contentType: "image/png",
-				},
-			],
-		};
-
-		const attachmentResponse =
-			await this.confluenceClient.contentAttachments.createOrUpdateAttachments(
-				attachmentDetails
-			);
-
-		return {
-			filename: uploadFilename,
-			id: attachmentResponse.results[0].extensions.fileId,
-			collection: `contentId-${attachmentResponse.results[0].container.id}`,
-		};
-	}
-
-	async uploadFile(
-		pageId: string,
-		pageFilePath: string,
-		fileNameToUpload: string,
-		currentAttachments: Record<
-			string,
-			{ filehash: string; attachmentId: string; collectionName: string }
-		>
-	): Promise<UploadedImageData | null> {
-		const testing = await this.adaptor.readBinary(
-			fileNameToUpload,
-			pageFilePath
-		);
-		if (testing) {
-			const spark = new SparkMD5.ArrayBuffer();
-			const currentFileMd5 = spark.append(testing.contents).end();
-			const pathMd5 = SparkMD5.hash(testing.filePath);
-			const uploadFilename = `${pathMd5}-${testing.filename}`;
-
-			if (
-				!!currentAttachments[uploadFilename] &&
-				currentAttachments[uploadFilename].filehash === currentFileMd5
-			) {
-				return {
-					filename: fileNameToUpload,
-					id: currentAttachments[uploadFilename].attachmentId,
-					collection:
-						currentAttachments[uploadFilename].collectionName,
-				};
-			}
-
-			const attachmentDetails = {
-				id: pageId,
-				attachments: [
-					{
-						file: Buffer.from(testing.contents),
-						filename: uploadFilename,
-						minorEdit: false,
-						comment: currentFileMd5,
-					},
-				],
-			};
-
-			const attachmentResponse =
-				await this.confluenceClient.contentAttachments.createOrUpdateAttachments(
-					attachmentDetails
-				);
-
-			return {
-				filename: fileNameToUpload,
-				id: attachmentResponse.results[0].extensions.fileId,
-				collection: `contentId-${attachmentResponse.results[0].container.id}`,
-			};
-		}
-
-		return null;
-	}
 }
-
-function getMermaidFileName(mermaidContent: string | undefined) {
-	const mermaidText = mermaidContent ?? "flowchart LR\nid1[Missing Chart]";
-	const pathMd5 = SparkMD5.hash(mermaidText);
-	const uploadFilename = `RenderedMermaidChart-${pathMd5}.png`;
-	return { uploadFilename, mermaidText };
-}
-
-interface UploadedImageData {
-	filename: string;
-	id: string;
-	collection: string;
-}
-
-interface TreeNode {
-	name: string;
-	children: TreeNode[];
-	file?: AdfFile;
-}
-const findCommonPath = (paths: string[]): string => {
-	const [firstPath, ...rest] = paths;
-	const commonPathParts = firstPath.split(path.sep);
-
-	rest.forEach((filePath) => {
-		const pathParts = filePath.split(path.sep);
-		for (let i = 0; i < commonPathParts.length; i++) {
-			if (pathParts[i] !== commonPathParts[i]) {
-				commonPathParts.splice(i);
-				break;
-			}
-		}
-	});
-
-	return commonPathParts.join(path.sep);
-};
-
-const createTreeNode = (name: string): TreeNode => ({
-	name,
-	children: [],
-});
-
-const addFileToTree = (
-	treeNode: TreeNode,
-	file: MarkdownFile,
-	relativePath: string,
-	fileNames: Set<string>
-) => {
-	const [folderName, ...remainingPath] = relativePath.split(path.sep);
-
-	if (remainingPath.length === 0) {
-		if (fileNames.has(file.fileName)) {
-			throw new Error(
-				`File name "${file.fileName}" is not unique across all folders.`
-			);
-		}
-		fileNames.add(file.fileName);
-		const adfFile = mdToADFConverter.convertMDtoADF(file);
-		treeNode.children.push({
-			...createTreeNode(folderName),
-			file: adfFile,
-		});
-	} else {
-		let childNode = treeNode.children.find(
-			(node) => node.name === folderName
-		);
-
-		if (!childNode) {
-			childNode = createTreeNode(folderName);
-			treeNode.children.push(childNode);
-		}
-
-		addFileToTree(childNode, file, remainingPath.join(path.sep), fileNames);
-	}
-};
-
-const createFolderStructure = (markdownFiles: MarkdownFile[]): TreeNode => {
-	const commonPath = findCommonPath(
-		markdownFiles.map((file) => file.absoluteFilePath)
-	);
-	const rootNode = createTreeNode(commonPath);
-	const fileNames = new Set<string>();
-
-	markdownFiles.forEach((file) => {
-		const relativePath = path.relative(commonPath, file.absoluteFilePath);
-		addFileToTree(rootNode, file, relativePath, fileNames);
-	});
-	console.log({ folderRoot: rootNode });
-	const processNode = (node: TreeNode) => {
-		if (!node.file) {
-			const indexFile = node.children.find(
-				(child) => path.parse(child.name).name === node.name
-			);
-
-			if (indexFile) {
-				node.file = indexFile.file;
-				node.children = node.children.filter(
-					(child) => child !== indexFile
-				);
-			} else {
-				node.file = {
-					folderName: node.name,
-					absoluteFilePath: path.join(commonPath, node.name),
-					fileName: `${node.name}.md`,
-					contents: FolderFile as JSONDocNode,
-					pageTitle: node.name,
-					frontmatter: {},
-					tags: [],
-					pageId: undefined,
-					dontChangeParentPageId: false,
-				};
-			}
-		}
-
-		node.children.forEach(processNode);
-	};
-
-	processNode(rootNode);
-
-	return rootNode;
-};
