--- src/mermaid_renderers/types.ts
@@ -1,8 +1,0 @@
-export interface ChartData {
-	name: string;
-	data: string;
-}
-
-export interface MermaidRenderer {
-	captureMermaidCharts(charts: ChartData[]): Promise<Map<string, Buffer>>;
-}
